requirements.txt
boto3==1.8.5
botocore==1.11.8
docutils==0.14
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.0
python-dateutil==2.7.3
PyYAML==3.13
s3transfer==0.1.13
six==1.11.0
urllib3==1.23
